Break patch operations into their own module
